requirements.txt
streamlit
requests
pandas
tldextract
psycopg2-binary
matplotlib
plotly